refactor: Update to import config helper function from helper module
